chore: setup tsdx
And remove microbundle
