--- .babelrc
@@ -1,16 +1,0 @@
-{
-  "presets": [
-    "@babel/preset-env",
-    "@babel/preset-typescript",
-    "@babel/preset-react"
-  ],
-  "env": {
-    "test": {
-      "presets": [
-        ["@babel/preset-env", { "targets": { "node": "current" } }],
-        "@babel/preset-typescript",
-        "@babel/preset-react"
-      ]
-    }
-  }
-}
