+++ package.json
@@ -21,22 +21,20 @@
     "utils"
   ],
   "scripts": {
-    "test": "jest",
-    "prebuild": "del dist",
-    "build": "microbundle --target web --no-sourcemap",
+    "start": "tsdx watch",
+    "build": "tsdx build",
     "postbuild": "del dist/**/*.{test,example}.d.ts",
-    "dev": "microbundle watch --target web --no-compress",
+    "test": "tsdx test",
     "docz:dev": "node scripts/create-index-doc.js && docz dev",
     "docz:build": "node scripts/create-index-doc.js && docz build",
     "lint": "eslint 'src/**/*.{ts,tsx}'",
     "create": "node scripts/create-hook.js",
     "semantic-release": "semantic-release"
   },
+  "dependencies": {
+    "use-subscription": "^1.3.0"
+  },
   "devDependencies": {
-    "@babel/core": "^7.8.4",
-    "@babel/preset-env": "^7.8.4",
-    "@babel/preset-react": "^7.8.3",
-    "@babel/preset-typescript": "^7.8.3",
     "@testing-library/jest-dom": "^5.1.1",
     "@testing-library/react": "^9.4.1",
     "@testing-library/user-event": "^10.0.0",
@@ -62,14 +60,14 @@
     "eslint-plugin-react": "7.18.3",
     "eslint-plugin-react-hooks": "2.x",
     "husky": "^4.2.3",
-    "jest": "^25.1.0",
     "lint-staged": "^10.0.7",
-    "microbundle": "^0.11.0",
     "prettier": "^1.19.1",
     "react": "16.12.0",
     "react-dom": "16.12.0",
     "semantic-release": "^17.0.4",
     "strman": "^2.0.1",
+    "tsdx": "^0.12.3",
+    "tslib": "^1.11.0",
     "typescript": "^3.8.2"
   },
   "peerDependencies": {
@@ -92,6 +90,9 @@
     }
   },
   "lint-staged": {
+    "js,ts,jsx,tsx": [
+      "eslint --fix"
+    ],
     "*.{js,ts,jsx,tsx,json,md,yml,html}": [
       "prettier --write"
     ]
@@ -101,8 +102,5 @@
     "singleQuote": true,
     "trailingComma": "all",
     "proseWrap": "always"
-  },
-  "dependencies": {
-    "use-subscription": "^1.3.0"
   }
 }
