+++ src/shared.ts
@@ -1,0 +1,6 @@
+export enum AsyncState {
+  initial = 'INITIAL',
+  pending = 'PENDING',
+  fullfilled = 'FULLFILLED',
+  rejected = 'REJECTED',
+}
