+++ src/use-promise/use-promise.test.tsx
@@ -1,6 +1,7 @@
 import * as React from 'react';
 import { render, cleanup, waitForElement } from '@testing-library/react';
-import { usePromise, AsyncState } from '../use-promise';
+import { usePromise } from '../use-promise';
+import { AsyncState } from '../shared';
 
 /**
  *
